+++ src/common/js/index.js
@@ -2,6 +2,7 @@
  * pwix:app-edit/src/common/js/index.js
  */
 
+import '../classes/display-unit.class.js';
 import '../classes/run-context.class.js';
 
 import { IAppEditable } from '../interfaces/iapp-editable.iface.js';
@@ -11,6 +12,7 @@ import './constants.js';
 import './configure.js';
 import './trace.js';
 //
+import './functions.js';
 import './i18n.js';
 
 AppEdit.IAppEditable = IAppEditable;