+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - Move classes and interfaces definition to common code, thus bumping minor candidate version number
     - IAppEditable methods are now prefixed with iAppEditable
+    - Extends AppPages.DisplayUnit class to define editionAllowed() method
 
 ### 1.0.1